requirements.txt
numpy~=1.26.2
matplotlib~=3.8.2
Pillow~=10.1.0